+++ pkgs/google_generative_ai/lib/src/chat.dart
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import 'dart:async';
+
 import 'api.dart';
 import 'content.dart';
 import 'model.dart';
@@ -64,7 +66,7 @@ final class ChatSession {
   Future<GenerateContentResponse> sendMessage(Content message) async {
     final lock = await _mutex.acquire();
     try {
-      final response = await _generateContent([..._history, message],
+      final response = await _generateContent(_history.followedBy([message]),
           safetySettings: _safetySettings, generationConfig: _generationConfig);
       if (response.candidates case [final candidate, ...]) {
         _history.add(message);
@@ -95,13 +97,10 @@ final class ChatSession {
   /// Waits to read the entire streamed response before recording the message
   /// and response and allowing pending messages to be sent.
   Stream<GenerateContentResponse> sendMessageStream(Content message) {
-    final mutexLock = _mutex.acquire(); // Acquire lock synchronously.
-    // TODO: Eagerly listen to the response and append to history, even if the
-    // returned stream doesn't have a listener.
-    return () async* {
-      final lock = await mutexLock;
+    final controller = StreamController<GenerateContentResponse>(sync: true);
+    _mutex.acquire().then((lock) async {
       try {
-        final responses = _generateContentStream([..._history, message],
+        final responses = _generateContentStream(_history.followedBy([message]),
             safetySettings: _safetySettings,
             generationConfig: _generationConfig);
         final content = <Content>[];
@@ -109,16 +108,19 @@ final class ChatSession {
           if (response.candidates case [final candidate, ...]) {
             content.add(candidate.content);
           }
-          yield response;
+          controller.add(response);
         }
         if (content.isNotEmpty) {
           _history.add(message);
           _history.add(_aggregate(content));
         }
-      } finally {
-        lock.release();
+      } catch (e, s) {
+        controller.addError(e, s);
       }
-    }();
+      lock.release();
+      unawaited(controller.close());
+    });
+    return controller.stream;
   }
 
   /// Aggregates a list of [Content] responses into a single [Content].
@@ -130,19 +132,28 @@ final class ChatSession {
     assert(contents.isNotEmpty);
     final role = contents.first.role ?? 'model';
     final textBuffer = StringBuffer();
+    // If non-null, only a single text part has been seen.
+    TextPart? previousText;
     final parts = <Part>[];
     void addBufferedText() {
-      if (textBuffer.isNotEmpty) {
+      if (textBuffer.isEmpty) return;
+      if (previousText case var singleText?) {
+        parts.add(singleText);
+        previousText = null;
+      } else {
         parts.add(TextPart(textBuffer.toString()));
-        textBuffer.clear();
       }
+      textBuffer.clear();
     }
 
     for (final content in contents) {
       for (final part in content.parts) {
         switch (part) {
           case TextPart(:final text):
-            textBuffer.write(text);
+            if (text.isNotEmpty) {
+              previousText = textBuffer.isEmpty ? part : null;
+              textBuffer.write(text);
+            }
           case DataPart():
             addBufferedText();
             parts.add(part);