+++ pkgs/google_generative_ai/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1-wip
 
 - Add additional API documentation.
+- Update the getting started instructions in the readme.
 
 ## 0.1.0